(feat: web3) implemented the interface for the MockSC, and made it pass the original tests.
