+++ lib/smartcontracts/__init__.py
@@ -1,0 +1,1 @@
+from lib.smartcontracts.MockSC import MockSC
